pyproject.toml
[project]
name = "simple-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "streamlit>=1.52.2",
    "watchdog>=6.0.0",
]
